+++ app/views/api/capabilities/show.builder
